requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

deep-sort-realtime>=1.3.2

opencv-python>=4.8.0
Pillow>=10.0.0

transformers>=4.36.0
sentence-transformers>=2.2.2

faiss-cpu>=1.7.4
langchain-huggingface==1.2.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0

fastapi>=0.110.0
uvicorn[standard]>=0.27.0

prometheus-client>=0.19.0

streamlit>=1.31.0
requests>=2.31.0

numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.66.0
